Change nMaxStakeAge back to unsigned int and put back in typecasting in wallet.cpp and kernel.cpp

+++ src/kernel.cpp
@@ -17,7 +17,7 @@ int64_t GetWeight(const int64_t &nIntervalBeginning, const int64_t &nIntervalEnd
     // this change increases active coins participating the hash and helps
     // to secure the network when proof-of-stake difficulty is low
 
-    return min(nIntervalEnd - nIntervalBeginning - nStakeMinAge, nStakeMaxAge);
+    return min(nIntervalEnd - nIntervalBeginning - nStakeMinAge, (int64_t)nStakeMaxAge);
 }
 
 // Get the last stake modifier and its generation time from a given block
