+++ src/main.h
@@ -76,7 +76,7 @@ extern CBlockIndex* pindexGenesisBlock;
 extern uint nTargetSpacing;
 extern uint nTargetStakeSpacing;
 extern uint nStakeMinAge;
-extern int64_t nStakeMaxAge;
+extern uint nStakeMaxAge;
 extern uint nNodeLifespan;
 extern int nCoinbaseMaturity;
 extern int nBestHeight;
